requirements.txt
langchain
transformers
chromadb
langchain-huggingface-hub
sentence-transformers
streamlit
